+++ app/src/main/java/jp/kentan/studentportalplus/data/component/MyClass.kt
@@ -2,8 +2,8 @@ package jp.kentan.studentportalplus.data.component
 
 
 data class MyClass(
-        val id          : Int = -1,
-        val hash        : Int,           // Murmur3.hash32(week + period + scheduleCode + credit + category + subject + instructor)
+        val id          : Long = -1,
+        val hash        : Long,          // Murmur3.hash64(week + period + scheduleCode + credit + category + subject + instructor)
         val week        : ClassWeekType, // 週
         val period      : Int,           // 時限
         val scheduleCode: String,        // 時間割コード
